# openapi.yaml
openapi: 3.0.1
info:
  title: ChatGPT Memory Plugin
  description: >
    A plugin that allows ChatGPT to have long-term memory by storing and retrieving conversation history.
    This plugin enables ChatGPT to recall specific exchanges and facts from previous interactions.
    To store an interaction, use the 'addMemory' endpoint with the user input and ChatGPT response.
    To retrieve relevant interactions, use the 'getMemory' endpoint with a query text.
    To delete a specific interaction, use the 'deleteMemory' endpoint with the index of the interaction to delete.
    Note: User input is limited to 1000 characters.

    The interactions can be tagged with predefined categories to help organize and filter the memory. The predefined categories are:
    - Personal: Personal information about the user.
    - Work: Work-related information and tasks.
    - Education: Information related to education and learning.
    - Entertainment: Information related to entertainment and leisure activities.
    - Other: Any other information that does not fit into the above categories.

    Recommendations:
    - When using the 'addMemory' endpoint, consider sending requests with smaller conversation pieces or a summary of the conversation to avoid exceeding the character limit.
    - If a request is too large or formatting takes too long, consider breaking it down into smaller requests.
    - Regularly check the conversation history size and consider deleting less relevant interactions to improve performance.
    - For the 'author' field, use 'ChatGPT' when storing responses generated by ChatGPT, and use 'User' when storing user inputs. This will ensure consistency in the authorship of interactions.

  version: 'v4'
servers:
  - url: http://localhost:5004
paths:
  /memory/{username}:
    get:
      operationId: getMemory
      summary: Get relevant interactions from the conversation history for the specified user
      description: >
        Retrieves relevant interactions from the conversation history for the specified user based on the query text and their cosine similarity with the query.
        The history includes user inputs, ChatGPT responses, embeddings, and timestamps for each interaction.
      parameters:
      - in: path
        name: username
        schema:
            type: string
        required: true
        description: The name of the user.
      - in: query
        name: query_text
        schema:
            type: string
        required: true
        description: The query text used to search for relevant interactions.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getMemoryResponse'
    post:
      operationId: addMemory
      summary: Add a new interaction to the conversation history
      description: >
        Adds a new interaction to the conversation history for the specified user.
        Requires the user input and the ChatGPT response to be provided.
        The user input is limited to 1000 characters.
        For the 'author' field, use 'ChatGPT' for responses generated by ChatGPT, and 'User' for user inputs.
      parameters:
      - in: path
        name: username
        schema:
            type: string
        required: true
        description: The name of the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/addMemoryRequest'
      responses:
        "200":
          description: OK
    delete:
      operationId: deleteMemory
      summary: Delete an interaction from the conversation history
      description: >
        Deletes a specific interaction from the conversation history for the specified user.
        Requires the index of the interaction to delete.
      parameters:
      - in: path
        name: username
        schema:
            type: string
        required: true
        description: The name of the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/deleteMemoryRequest'
      responses:
        "200":
          description: OK
components:
  schemas:
    getMemoryResponse:  
      type: object
      properties:
        history:
          type: array
          items:
            type: object
            properties:
              user_input:
                type: string
                description: The input provided by the user.
              chatgpt_response:
                type: string
                description: The response provided by ChatGPT.
              timestamp:
                type: string
                format: date-time
                description: The timestamp of the interaction.
          description: The conversation history.
    addMemoryRequest:
      type: object
      required:
      - user_input
      - chatgpt_response
      - document_name
      - author
      properties:
        user_input:
          type: string
          description: The input provided by the user (limited to 1000 characters).
          required: true
          maxLength: 1000
        chatgpt_response:
          type: string
          description: The response provided by ChatGPT.
          required: true
        document_name:
          type: string
          description: The name of the document being stored.
          required: true
        author:
          type: string
          description: The author of the content. Use 'ChatGPT' for responses generated by ChatGPT, and 'User' for user inputs.
          required: true
        category:
          type: string
          description: The category used to tag or categorize the interaction. Choose from 'Personal', 'Work', 'Education', 'Entertainment', or 'Other'.
          enum: [Personal, Work, Education, Entertainment, Other]
    deleteMemoryRequest:
      type: object
      required:
      - interaction_idx
      properties:
        interaction_idx:
          type: integer
          description: The index of the interaction to delete.
          required: true